refactor(home): split HomePage orchestration into dedicated composables
